Update YAMLs, Tests
Updated the base YAML regex file, commented out failing tests which are
failing due to my preference to not implement the jsParseBits portion
of the original Python library.

If I ever get clarification on the code, I can always go back and
modify. For now, the test files will be testing that the Julia code
hasn’t broken.